+++ CHANGELOG.md
@@ -1,6 +1,46 @@
 # Changelog
 
+## Next moves
+### Changed
+  * [Deprecation] Calling `restClientSdk.foo.xxx` is deprecated, you must now call `restClientSdk.getRepository('foo').xxx`
+
 ## Unreleased
+
+### Changed
+  * [Breaking] RestClientSdk now takes a Mapping instance instead of a clientList. This instance is required
+  * [Breaking] `getName` has been removed and replaced by the classmetadata key. Its return was previously sent to the serializer, it's now the key attribute from the mapping that is sent now.
+  Beware that now, the `key` is now used both to call the repository name, and passed to the serializer
+  Before:
+  ```js
+  class SomeClient extends AbstractClient {
+      getName() {
+          return 'SomeClient';
+      }
+  }
+
+  new RestClientSdk(
+      // ...
+      {
+        foo: SomeClient,
+      }
+  )
+  ```
+
+  Now:
+  ```js
+    const metadata = new ClassMetadata('foo', 'some_endpoint');
+    mapping = new Mapping();
+    mapping.setMapping([ metadata ]);
+
+    const sdk = new RestClientSdk(/* ... */, mapping); // `foo` will be used in serializer too
+    sdk.foo.find();
+  ```
+
+
+### Removed
+  * AbstractClient does not need to specify `getPathBase` anymore: this is generated via the classmetata. You can still override it.
+  * `getEntityURI` is not mandatory anymore. it will be autogenerate by the AbstractClient. You can still override it
+  * There is no need to create an empty client extending `AbstractClient` now, the default will be `AbstractClient`
 
 ## [1.3.2] - 2017-12-04